+++ script/update-doc-repo.pl
@@ -22,6 +22,7 @@ use version;
 use vars qw(
 	$HELP
 	$DEBUG
+	$SKIP_INDEX
 
 	$BOOTSTRAPVERSION
 	$DESCRIPTIONS
@@ -42,6 +43,7 @@ use vars qw(
 );
 
 $DEBUG = 0;
+$SKIP_INDEX = 0;
 $ROOT = '/mnt/docs.opennms.org';
 
 $BOOTSTRAPVERSION = '3.3.4';
@@ -71,11 +73,12 @@ $DESCRIPTIONS = {
 };
 
 my $result = GetOptions(
-	"h|help"     => \$HELP,
-	"d|debug"    => \$DEBUG,
+	"h|help"       => \$HELP,
+	"d|debug"      => \$DEBUG,
+	"s|skip-index" => \$SKIP_INDEX,
 
-	"r|root=s"   => \$ROOT,
-	"b|branch=s" => \$BRANCH,
+	"r|root=s"     => \$ROOT,
+	"b|branch=s"   => \$BRANCH,
 );
 
 if (not defined $ROOT or $ROOT eq "") {
@@ -127,31 +130,47 @@ if (not defined $BRANCH or $BRANCH eq "") {
 
 my $lockfile = File::Spec->catfile($ROOT, '.update-doc-repo.lock');
 
-print "* waiting for lock...\n";
 ### set up a lock - lasts until object loses scope
-if (my $lock = new File::NFSLock {
-	file      => $lockfile,
-	lock_type => LOCK_EX,
-	blocking_timeout   => 60 * 60, # 60 minutes
-	stale_lock_timeout => 60 * 60 * 2, # 2 hours
-}) {
-	# update the lock file
-	open(FILE, ">$lockfile") || die "Failed to lock $ROOT: $!\n";
-	print FILE localtime(time);
-	$lock->uncache;
-	print "* got lock -- updating docs.\n";
+my $lock;
+my $timeout = time() + (60 * 60); # 60 minutes
+
+LOCK: while(time() < $timeout) {
+	do_log("- waiting for lock...");
+
+	$lock = new File::NFSLock {
+		file      => $lockfile,
+		lock_type => LOCK_EX|LOCK_NB,
+		blocking_timeout   => $timeout,
+		stale_lock_timeout => $timeout * 2,
+	};
+
+	# if we get a lock, update the lock file
+	if ($lock) {
+		open(FILE, ">$lockfile") || die "Failed to lock $ROOT: $!\n";
+		print FILE localtime(time());
+		$lock->uncache;
+		do_log("- got lock -- updating documentation");
+		last LOCK;
+	}
+
+	# otherwise keep waiting
+	sleep(5);
+}
+
+if (!$lock) {
+	die "Couldn't lock $lockfile [$File::NFSLock::errstr]";
+}
 
 ##### START UPDATING, INSIDE LOCK #####
 
-print "- Creating document directory '$INSTALLDIR'... ";
+do_log("- Creating document directory '$INSTALLDIR'");
 mkpath($INSTALLDIR);
-print "done\n";
 
 if (-f $DOCS) {
 	$DOCDIR = tempdir( CLEANUP => 1 );
-	print STDERR "! Created temporary directory $DOCDIR.\n" if ($DEBUG);
+	do_debug("! Created temporary directory $DOCDIR");
 	mkpath($DOCDIR);
-	print "- Unpacking '$DOCS' into temporary directory... ";
+	do_log("- Unpacking '$DOCS' into temporary directory");
 	chdir($DOCDIR);
 	if ($DOCS =~ /\.(jar|zip)$/) {
 		system('unzip', '-q', $DOCS) == 0 or die "Failed to unpack $DOCS into $DOCDIR: $!\n";
@@ -161,7 +180,6 @@ if (-f $DOCS) {
 		die "Unhandled file: $DOCS\n";
 	}
 	chdir($INSTALLDIR);
-	print "done\n";
 } else {
 	$DOCDIR = $DOCS;
 }
@@ -196,34 +214,47 @@ open (FILEOUT, '>', $versionfile) or die "Failed to open $versionfile for writin
 print FILEOUT $VERSION;
 close(FILEOUT) or die "Failed to close $versionfile: $!\n";
 
-@PROJECTS = get_projects($ROOT);
-update_indexes();
-create_release_symlinks();
-fix_permissions($INSTALLDIR);
+if (not $SKIP_INDEX) {
+	@PROJECTS = get_projects($ROOT);
+	update_indexes();
+	create_release_symlinks();
+	fix_permissions($INSTALLDIR);
+}
 
-##### FINISHED UPDATING, FINISH LOCK #####
+END {
+	##### FINISHED UPDATING, CLOSE LOCK #####
 
-	unlink($lockfile) or die "Failed to remove $lockfile: $!\n";
-	close(FILE) or die "Failed to close $lockfile: $!\n";
-	$lock->unlock();
+	if (defined $lockfile and defined $lock) {
+		do_log("- cleaning up lock...");
+		unlink($lockfile) or die "Failed to remove $lockfile: $!\n";
+		close(FILE) or die "Failed to close $lockfile: $!\n";
+		$lock->unlock();
+	}
+}
 
-	exit 0;
-} else {
-	die "Couldn't lock $lockfile [$File::NFSLock::errstr]";
+exit 0;
+
+sub do_log {
+	print localtime(time()) . " " . join('', @_) . "\n";
+}
+
+sub do_debug {
+	return unless $DEBUG;
+	print localtime(time()) . " " . join('', @_) . "\n";
 }
 
 sub get_projects {
 	my $projectsroot = shift;
 
-	print STDERR "! Getting projects from $projectsroot:\n" if ($DEBUG);
+	do_debug("! Getting projects from $projectsroot:");
 	my $projects = {};
 	opendir(DIR, $projectsroot) or die "Failed to open $projectsroot for reading: $!\n";
 	while (my $entry = readdir(DIR)) {
-		my $path = File::Spec->catdir($projectsroot, $entry);
 		next if ($entry =~ /^\./);
 		next if ($entry =~ /^\@eaDir/);
 		next if ($entry =~ /^index\.html$/);
 		next if ($entry =~ /^(api|documentation|Minion-Events|OpenNMS|PRIS|SMNnepO)$/);
+		my $path = File::Spec->catdir($projectsroot, $entry);
 		next if (-l $path);
 		next unless (-d $path);
 
@@ -233,7 +264,7 @@ sub get_projects {
 			path        => $path,
 		};
 
-		print STDERR "! * Found: ", $project->{'description'}, "\n" if ($DEBUG);
+		do_debug("! * Found: ", $project->{'description'});
 		$projects->{$entry} = $project;
 	}
 	closedir(DIR) or die "Failed to close $projectsroot: $!\n";
@@ -262,25 +293,24 @@ sub get_releases {
 	my $project = shift;
 	my @releases;
 
-	print STDERR "! Getting release types for project " . $project->{'description'} . "\n" if ($DEBUG);
+	do_debug("! Getting release types for project " . $project->{'description'});
 	opendir(DIR, $project->{'path'}) or die "Failed to open " . $project->{'path'} . " for reading: $!\n";
 	for my $entry (sort readdir(DIR)) {
-		my $path = File::Spec->catdir($project->{'path'}, $entry);
 		next unless ($entry =~ /^(branches|releases)$/);
-		print STDERR "! * Found release type: $entry\n" if ($DEBUG);
-		print STDERR "! * Searching for releases:\n" if ($DEBUG);
+		do_debug("! * Found release type: $entry");
+		do_debug("! * Searching for releases:");
 
 		my $releasesdir = File::Spec->catdir($project->{'path'}, $entry);
 		opendir(RELEASESDIR, $releasesdir) or die "Failed to open $releasesdir for reading: $!\n";
 		for my $releaseentry (sort readdir(RELEASESDIR)) {
-			my $releasedir = File::Spec->catdir($releasesdir, $releaseentry);
 			next if ($releaseentry =~ /^\./);
 			next if ($releaseentry =~ /^\@eaDir/);
 			next if ($releaseentry =~ /^index\.html$/);
+			my $releasedir = File::Spec->catdir($releasesdir, $releaseentry);
 			next if (-l $releasedir);
 			next unless (-d $releasedir);
 
-			print STDERR "!   * Found: $releaseentry\n" if ($DEBUG);
+			do_debug("!   * Found: $releaseentry");
 
 			my $release = {
 				type    => $entry,
@@ -318,17 +348,17 @@ sub get_docs_for_release {
 	my $docs = {};
 
 	my $display = $release->{'type'} . '/' . $release->{'name'};
-	print STDERR "! Finding documentation in $display... " if ($DEBUG);
+	do_debug("! Finding documentation in $display...");
 	opendir(DIR, $release->{'path'}) or die "Failed to open " . $release->{'path'} . " for reading: $!\n";
 	for my $entry (sort readdir(DIR)) {
-		my $docpath = File::Spec->catdir($release->{'path'}, $entry);
 		next if ($entry =~ /^\./);
 		next if ($entry =~ /^\@eaDir/);
 		next if ($entry =~ /^index\.html$/);
+		my $docpath = File::Spec->catdir($release->{'path'}, $entry);
 		next if (-l $docpath);
 		next unless (-d $docpath);
 
-		print STDERR "$entry... " if ($DEBUG);
+		do_debug("  $entry...");
 
 		my $doc = {
 			name => $entry,
@@ -354,7 +384,6 @@ sub get_docs_for_release {
 
 		$docs->{$entry} = $doc;
 	}
-	print STDERR "done\n" if ($DEBUG);
 
 	return $docs;
 }
@@ -373,13 +402,17 @@ sub get_branches_for_project {
 }
 
 sub update_indexes {
+	do_log("- updating indexes...");
 
 	opendir(DIR, $ROOT) or die "Failed to open $ROOT for reading: $!\n";
 
 	my $roottext = "<h3>OpenNMS Projects</h3>\n<ul>\n";
 	for my $project (@PROJECTS) {
+
 		my $desc = $project->{'description'};
 		my $projectdir = $project->{'path'};
+
+		do_debug("- updating indexes for " . $desc);
 
 		$roottext .= "	<li class=\"project\">\n";
 		$roottext .= get_link($desc, File::Spec->catdir($projectdir, 'index.html'), $ROOT) . "\n";
@@ -430,7 +463,7 @@ sub create_release_symlinks {
 			if (-e $latestfile) {
 				unlink $latestfile;
 			}
-			print STDERR '! ln -sf ' . $release->{'name'} . ' ' . $latestfile . "\n" if ($DEBUG);
+			do_debug('! ln -sf ' . $release->{'name'} . ' ' . $latestfile);
 			symlink($release->{'name'}, $latestfile);
 		}
 	}
@@ -795,7 +828,7 @@ sub copy_doc_directory {
 
 	my $to = File::Spec->catdir($INSTALLDIR, $guide);
 
-	print "- Copying $guide to '$to'... ";
+	do_log("- Copying $guide to '$to'");
 	find({
 		wanted => sub {
 			return unless (-f $File::Find::name);
@@ -812,7 +845,7 @@ sub copy_doc_directory {
 				#system('chmod', '755', $dirname) == 0 or die "Failed to fix ownership on $dirname: $!\n";
 			}
 
-			#print "- copy: $fromfile -> $tofile\n";
+			do_debug("  - copy: $fromfile -> $tofile");
 			copy($fromfile, $tofile) or die "Failed to copy '$fromfile' to '$tofile': $!\n";
 			#system('chmod', '644', $tofile) == 0 or die "Failed to fix ownership on $tofile: $!\n";
 
@@ -824,7 +857,6 @@ sub copy_doc_directory {
 		follow => 1,
 		no_chdir => 1,
 	}, $from);
-	print "done\n";
 }
 
 sub process_docbook_docdir {
@@ -863,6 +895,7 @@ sub process_docbook_docdir {
 					#system('chmod', '755', $dirname) == 0 or die "Failed to fix ownership on $dirname: $!\n";
 				}
 
+				do_debug("  - copy: $fromfile -> $tofile");
 				copy($fromfile, $tofile) or die "Failed to copy '$fromfile' to '$tofile': $!\n";
 				#system('chmod', '644', $tofile) == 0 or die "Failed to fix ownership on $tofile: $!\n";
 			},
@@ -880,7 +913,7 @@ sub process_docbook_docdir {
 		my ($name) = $file =~ /^(.*?)\.(pdf|html)$/;
 		next if ($done->{$name});
 
-		print "- Processing $name... ";
+		do_debug("- Processing $name");
 
 		my $mappedname = (exists $mapping->{$name}? $mapping->{$name} : $name);
 		my $to = File::Spec->catdir($INSTALLDIR, $mappedname);
@@ -897,6 +930,7 @@ sub process_docbook_docdir {
 				#system('chmod', '755', $dirname) == 0 or die "Failed to fix ownership on $dirname: $!\n";
 			}
 
+			do_debug("  - copy: $fromfile -> $tofile");
 			copy($fromfile, $tofile) or die "Failed to copy '$fromfile' to '$tofile': $!\n";
 			#system('chmod', '644', $tofile) or die "Failed to fix ownership on $tofile: $!\n";
 
@@ -907,7 +941,6 @@ sub process_docbook_docdir {
 			}
 		}
 
-		print "done\n";
 		$done->{$name}++;
 	}
 }
@@ -917,7 +950,7 @@ sub process_javadoc_docdir {
 	my $to   = File::Spec->catdir($INSTALLDIR, 'javadoc');
 	mkpath($to) unless (-d $to);
 
-	print "- Copying javadoc to '$to'... ";
+	do_log("- Copying javadoc to '$to'");
 	system('rsync', '-r', '--delete', $from.'/', $to.'/') == 0 or die "Failed to sync from $from to $to: $!\n";
 	print "done\n";
 }
